SNOW-1296757: local testing fix math.nan related issue (#1498)
