+++ CHANGELOG.md
@@ -7,6 +7,8 @@
 #### Bug Fixes
 
 - Fixed a bug in DataFrameReader.csv unable to handle quoted values containing delimiter.
+- Fixed a bug that when there is `None` value in arithmetic calculation, the output should remain `None` instead of `math.nan`.
+- Fixed a bug in function `sum` and `covar_pop` that when there is `math.nan` in the data, the output should also be `math.nan`.
 
 ## 1.16.0 (TBD)
 
@@ -30,7 +32,6 @@
 
 - Fixed a bug that caused NaT and NaN values to not be recognized.
 - Fixed a bug when inferring schema, single quotes are added to stage files already have single quotes.
-
 
 ## 1.15.0 (2024-04-24)
